fix: average color
